EN:AGO:Post comment stat count

+++ CommentManagement/src/main/java/ago/app/comment/es/ElasticSearchQuery.java
@@ -1,5 +1,7 @@
 package ago.app.comment.es;
 
+import ago.app.comment.base.ReactionStat;
+import ago.app.comment.base.dto.PostCommentDTO;
 import ago.app.comment.base.vo.PostCommentVO;
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import co.elastic.clients.elasticsearch._types.Result;
@@ -24,6 +26,8 @@ public class ElasticSearchQuery {
     private ElasticsearchClient elasticsearchClient;
 
     private final String indexName = EsStatIndexConst.INDEX_post_comment;
+    private final String indexNamePostCommentStat = EsStatIndexConst.INDEX_post_comment_reaction_stat;
+
 
 
     public ErrorResponse createOrUpdateDocument(PostCommentVO postCommentVO, int savableStatus ) throws IOException {
@@ -34,6 +38,12 @@ public class ElasticSearchQuery {
             response = elasticsearchClient.index(i -> i.index(indexName)
                                 .document(postCommentVO)
               );
+            IndexResponse finalResponse = response;
+            elasticsearchClient.index(i -> i.index(indexNamePostCommentStat).id( finalResponse.id() )
+                    .document(new ReactionStat())
+            );
+
+            updatePostStatForComment( SavableConst.NEW, postCommentVO.getPostId() );
         }
         else if ( SavableConst.MODIFY == savableStatus)
         {
@@ -53,35 +63,58 @@ public class ElasticSearchQuery {
         return  new ErrorResponse<Result>( ErrorConstant.ERROR, "Error" );
     }
 
-    public PostCommentVO getDocumentById(String id) throws IOException {
-        PostCommentVO product = null;
-        GetResponse<PostCommentVO> response = elasticsearchClient.get(g -> g
+    public PostCommentDTO getDocumentById(String id) throws IOException {
+        PostCommentDTO product = null;
+        GetResponse<PostCommentDTO> response = elasticsearchClient.get(g -> g
                         .index(indexName)
                         .id(id),
-                PostCommentVO.class
+                PostCommentDTO.class
         );
 
         if (response.found()) {
             product = response.source();
             product.setCommentId(response.id());
+            product.setReactionStat(getReactionStatById(response.id()));
         }
-
         return product;
     }
 
     public String deleteDocumentById(String id) throws IOException {
 
-        DeleteRequest request = DeleteRequest.of(d -> d.index(indexName).id(id));
+        GetResponse<PostCommentDTO> response = elasticsearchClient.get(g -> g
+                        .index(indexName)
+                        .id(id),
+                PostCommentDTO.class );
+        if ( response.found()) {
 
-        DeleteResponse deleteResponse = elasticsearchClient.delete(request);
-        if (Objects.nonNull(deleteResponse.result()) && !deleteResponse.result().name().equals("NotFound")) {
-            return new StringBuilder("Product with id " + deleteResponse.id() + " has been deleted.").toString();
+            DeleteRequest request = DeleteRequest.of(d -> d.index(indexName).id(id));
+
+            DeleteResponse deleteResponse = elasticsearchClient.delete(request);
+            if (Objects.nonNull(deleteResponse.result()) && !deleteResponse.result().name().equals("NotFound")) {
+
+                // update post stat
+                updatePostStatForComment( SavableConst.DELETE, response.source().getPostId() );
+
+                return new StringBuilder("Product with id " + deleteResponse.id() + " has been deleted.").toString();
+            }
         }
-        System.out.println("Product not found");
-        return new StringBuilder("Product with id " + deleteResponse.id() + " does not exist.").toString();
+        return new StringBuilder("Product with id " + id + " does not exist.").toString();
 
     }
+    public ReactionStat getReactionStatById(String id) throws IOException {
+        ReactionStat product = null;
+        GetResponse<ReactionStat> response = elasticsearchClient.get(g -> g
+                        .index(indexNamePostCommentStat)
+                        .id(id),
+                ReactionStat.class
+        );
 
+        if (response.found()) {
+            product = response.source();
+        }
+
+        return product;
+    }
     public List<PostCommentVO> searchAllDocuments() throws IOException {
 
         SearchRequest searchRequest = SearchRequest.of(s -> s.index(indexName));
@@ -95,5 +128,32 @@ public class ElasticSearchQuery {
 
         }
         return products;
+    }
+    public void updatePostStatForComment( int savableStatus, String id ) throws IOException
+    {
+        if ( SavableConst.NEW == savableStatus )
+        {
+            elasticsearchClient.update( u -> u
+                            .index(EsStatIndexConst.INDEX_post_reaction_stat)
+                            .id(id)
+                            .script(s -> s
+                                    .inline(i -> i
+                                            .source("ctx._source.totalComment += 1")
+                                    )
+                            ),
+                    Void.class);
+        }
+        else if ( SavableConst.DELETE == savableStatus )
+        {
+            elasticsearchClient.update( u -> u
+                            .index(EsStatIndexConst.INDEX_post_reaction_stat)
+                            .id(id)
+                            .script(s -> s
+                                    .inline(i -> i
+                                            .source("ctx._source.totalComment -= 1")
+                                    )
+                            ),
+                    Void.class);
+        }
     }
 }
